fix: preserve full chain IDs in browser bindings (#832)

+++ helios-ts/src/linea.rs
@@ -93,8 +93,8 @@ impl LineaClient {
     }
 
     #[wasm_bindgen]
-    pub fn chain_id(&self) -> u32 {
-        self.chain_id as u32
+    pub fn chain_id(&self) -> u64 {
+        self.chain_id
     }
 
     #[wasm_bindgen]
